ANT-361 : introducing inheritance for marginal cost output variable
As the new link cost variable (say "reduced cost") to add is very similar to marginal cost class, we want to make marginal cost a child
of a base class that is intended to be a parent of both marginal cost and reduced cost.
Reduced cost is not implemented here, but the inheritance is in place.